MINOR: Merge org.apache.kafka.server.common.OffsetAndEpoch and org.apache.kafka.raft.OffsetAndEpoch
