+++ metadata/src/main/java/org/apache/kafka/image/MetadataImage.java
@@ -20,7 +20,7 @@ package org.apache.kafka.image;
 import org.apache.kafka.image.node.MetadataImageNode;
 import org.apache.kafka.image.writer.ImageWriter;
 import org.apache.kafka.image.writer.ImageWriterOptions;
-import org.apache.kafka.raft.OffsetAndEpoch;
+import org.apache.kafka.server.common.OffsetAndEpoch;
 
 import java.util.Objects;
