+++ metadata/src/test/java/org/apache/kafka/controller/OffsetControlManagerTest.java
@@ -23,8 +23,8 @@ import org.apache.kafka.common.metadata.EndTransactionRecord;
 import org.apache.kafka.common.metadata.NoOpRecord;
 import org.apache.kafka.common.utils.LogContext;
 import org.apache.kafka.raft.Batch;
-import org.apache.kafka.raft.OffsetAndEpoch;
 import org.apache.kafka.server.common.ApiMessageAndVersion;
+import org.apache.kafka.server.common.OffsetAndEpoch;
 import org.apache.kafka.timeline.TrackingSnapshotRegistry;
 
 import org.junit.jupiter.api.Test;
