+++ raft/src/main/java/org/apache/kafka/raft/ValidOffsetAndEpoch.java
@@ -16,6 +16,8 @@
  */
 package org.apache.kafka.raft;
 
+import org.apache.kafka.server.common.OffsetAndEpoch;
+
 import java.util.Objects;
 
 public final class ValidOffsetAndEpoch {
