+++ raft/src/main/java/org/apache/kafka/snapshot/FileRawSnapshotReader.java
@@ -19,7 +19,7 @@ package org.apache.kafka.snapshot;
 import org.apache.kafka.common.record.FileRecords;
 import org.apache.kafka.common.record.Records;
 import org.apache.kafka.common.record.UnalignedRecords;
-import org.apache.kafka.raft.OffsetAndEpoch;
+import org.apache.kafka.server.common.OffsetAndEpoch;
 
 import java.io.IOException;
 import java.io.UncheckedIOException;
