+++ raft/src/main/java/org/apache/kafka/snapshot/SnapshotPath.java
@@ -16,7 +16,7 @@
  */
 package org.apache.kafka.snapshot;
 
-import org.apache.kafka.raft.OffsetAndEpoch;
+import org.apache.kafka.server.common.OffsetAndEpoch;
 
 import java.nio.file.Path;
