+++ raft/src/main/java/org/apache/kafka/snapshot/Snapshots.java
@@ -19,8 +19,8 @@ package org.apache.kafka.snapshot;
 import org.apache.kafka.common.utils.BufferSupplier;
 import org.apache.kafka.common.utils.Utils;
 import org.apache.kafka.raft.KafkaRaftClient;
-import org.apache.kafka.raft.OffsetAndEpoch;
 import org.apache.kafka.raft.internals.IdentitySerde;
+import org.apache.kafka.server.common.OffsetAndEpoch;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
