+++ raft/src/test/java/org/apache/kafka/snapshot/SnapshotsTest.java
@@ -17,7 +17,7 @@
 package org.apache.kafka.snapshot;
 
 import org.apache.kafka.common.utils.Utils;
-import org.apache.kafka.raft.OffsetAndEpoch;
+import org.apache.kafka.server.common.OffsetAndEpoch;
 import org.apache.kafka.test.TestUtils;
 
 import org.junit.jupiter.api.Test;
